fix: timestamp format includes timezone offset instead of UTC (#48)
- Implement SqliteConnection.ClearAllPools() to resolve 'file in use' errors
- Fix UTC timestamp normalization to ensure stable metrics across timezones
- Add proper database cleanup before and after test execution
- Address nullable reference warnings and follow xUnit best practices

Closes #48
