+++ minisoc.sln
@@ -1,10 +1,20 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
+﻿Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.2.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "server", "server", "{DAD03CFC-D5FA-AC94-41C7-9ABAB326F09E}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MiniSOC.Server", "server\MiniSOC.Server\MiniSOC.Server.csproj", "{23B7755A-B64A-D4AC-018A-CD26DED39CD0}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "agent", "agent", "{B2E3D4F5-A6B7-4C8D-9E0F-1A2B3C4D5E6F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MiniSOC.Agent", "agent\MiniSOC.Agent\MiniSOC.Agent.csproj", "{A1B2C3D4-E5F6-4A5B-8C9D-0E1F2A3B4C5D}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{D6BCD86B-908C-4AAB-A5DA-1B0336BAF01F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MiniSOC.Agent.Tests", "tests\MiniSOC.Agent.Tests\MiniSOC.Agent.Tests.csproj", "{D50F83ED-1E77-4462-BA09-97291915BD9F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MiniSOC.Server.Tests", "tests\MiniSOC.Server.Tests\MiniSOC.Server.Tests.csproj", "{C1D2E3F4-B5A6-47D8-9E0F-1A2B3C4D5E6F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -16,12 +26,27 @@ Global
 		{23B7755A-B64A-D4AC-018A-CD26DED39CD0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{23B7755A-B64A-D4AC-018A-CD26DED39CD0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{23B7755A-B64A-D4AC-018A-CD26DED39CD0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A1B2C3D4-E5F6-4A5B-8C9D-0E1F2A3B4C5D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1B2C3D4-E5F6-4A5B-8C9D-0E1F2A3B4C5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1B2C3D4-E5F6-4A5B-8C9D-0E1F2A3B4C5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1B2C3D4-E5F6-4A5B-8C9D-0E1F2A3B4C5D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D50F83ED-1E77-4462-BA09-97291915BD9F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D50F83ED-1E77-4462-BA09-97291915BD9F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D50F83ED-1E77-4462-BA09-97291915BD9F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D50F83ED-1E77-4462-BA09-97291915BD9F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C1D2E3F4-B5A6-47D8-9E0F-1A2B3C4D5E6F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C1D2E3F4-B5A6-47D8-9E0F-1A2B3C4D5E6F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C1D2E3F4-B5A6-47D8-9E0F-1A2B3C4D5E6F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C1D2E3F4-B5A6-47D8-9E0F-1A2B3C4D5E6F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{23B7755A-B64A-D4AC-018A-CD26DED39CD0} = {DAD03CFC-D5FA-AC94-41C7-9ABAB326F09E}
+		{A1B2C3D4-E5F6-4A5B-8C9D-0E1F2A3B4C5D} = {B2E3D4F5-A6B7-4C8D-9E0F-1A2B3C4D5E6F}
+		{D50F83ED-1E77-4462-BA09-97291915BD9F} = {D6BCD86B-908C-4AAB-A5DA-1B0336BAF01F}
+		{C1D2E3F4-B5A6-47D8-9E0F-1A2B3C4D5E6F} = {D6BCD86B-908C-4AAB-A5DA-1B0336BAF01F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6351CFB7-171B-4371-8B11-C6428B15F058}
