+++ tests/MiniSOC.Server.Tests/GlobalUsings.cs
@@ -1,0 +1,2 @@
+global using Xunit;
+global using Microsoft.Data.Sqlite;
