+++ tests/MiniSOC.Server.Tests/HealthEndpointTests.cs
@@ -30,11 +30,14 @@ public class HealthEndpointTests : IClassFixture<WebApplicationFactory<Program>>
         Assert.Equal(HttpStatusCode.OK, response.StatusCode);
 
         var json = await response.Content.ReadAsStringAsync();
-        var healthResponse = JsonSerializer.Deserialize<HealthResponse>(json, new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
+        var options = new JsonSerializerOptions { PropertyNameCaseInsensitive = true };
+        var healthResponse = JsonSerializer.Deserialize<HealthResponse>(json, options);
 
         Assert.NotNull(healthResponse);
         Assert.Equal("healthy", healthResponse.Status);
         Assert.Equal("0.1.0", healthResponse.Version);
+        
+        // Verify timestamp is within a reasonable range (last 2 minutes)
         Assert.InRange(healthResponse.Timestamp, DateTime.UtcNow.AddMinutes(-1), DateTime.UtcNow.AddMinutes(1));
     }
 }
